LayoutTests/accessibility/mac/native-text-control-set-selected-textmarker-range-expected.txt

Test for setting the selected TextMarkerRange in a native text control: textarea and text field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


selected text "This is a textarea."
selected text "This is a "
selected text " is a "
selected text "This is a text field."
selected text "This is a "
selected text " is a text"
PASS successfullyParsed is true

TEST COMPLETE
